+++ Source/SwiftLintFramework/Helpers/Exec.swift
@@ -1,0 +1,61 @@
+import Foundation
+
+/// Namespace for utilities to execute a child process.
+enum Exec {
+    /// The result of running the child process.
+    struct Results {
+        /// The process's exit status.
+        let terminationStatus: Int32
+        /// The data from stdout and optionally stderr.
+        let data: Data
+        /// The `data` reinterpreted as a string with whitespace trimmed; `nil` for the empty string.
+        var string: String? {
+            let encoded = String(data: data, encoding: .utf8) ?? ""
+            let trimmed = encoded.trimmingCharacters(in: .whitespacesAndNewlines)
+            return trimmed.isEmpty ? nil : trimmed
+        }
+    }
+
+    /**
+     Run a command with arguments and return its output and exit status.
+
+     - parameter command: Absolute path of the command to run.
+     - parameter arguments: Arguments to pass to the command.
+     - parameter currentDirectory: Current directory for the command.  By default
+                                   the parent process's current directory.
+     */
+     static func run(_ command: String,
+                     _ arguments: [String] = [],
+                     currentDirectory: String = FileManager.default.currentDirectoryPath) -> Results {
+        let process = Process()
+        process.arguments = arguments
+
+        let pipe = Pipe()
+        process.standardOutput = pipe
+
+        do {
+#if canImport(Darwin)
+            if #available(macOS 10.13, *) {
+                process.executableURL = URL(fileURLWithPath: command)
+                process.currentDirectoryURL = URL(fileURLWithPath: currentDirectory)
+                try process.run()
+            } else {
+                process.launchPath = command
+                process.currentDirectoryPath = currentDirectory
+                process.launch()
+            }
+#else
+            process.executableURL = URL(fileURLWithPath: command)
+            process.currentDirectoryURL = URL(fileURLWithPath: currentDirectory)
+            try process.run()
+#endif
+        } catch {
+            return Results(terminationStatus: -1, data: Data())
+        }
+
+        let file = pipe.fileHandleForReading
+        let data = file.readDataToEndOfFile()
+        process.waitUntilExit()
+        return Results(terminationStatus: process.terminationStatus, data: data)
+    }
+}