+++ Source/swiftlint/Helpers/LintableFilesVisitor.swift
@@ -48,13 +48,16 @@ struct LintableFilesVisitor {
     let cache: LinterCache?
     let parallel: Bool
     let allowZeroLintableFiles: Bool
+    let stableGitRevision: String?
+    let explicitConfigurationPaths: [String]
     let mode: LintOrAnalyzeModeWithCompilerArguments
     let block: (CollectedLinter) -> Void
 
     init(paths: [String], action: String, useSTDIN: Bool,
          quiet: Bool, useScriptInputFiles: Bool, forceExclude: Bool, useExcludingByPrefix: Bool,
          cache: LinterCache?, parallel: Bool,
-         allowZeroLintableFiles: Bool, block: @escaping (CollectedLinter) -> Void) {
+         allowZeroLintableFiles: Bool, stableGitRevision: String?, explicitConfigurationPaths: [String],
+         block: @escaping (CollectedLinter) -> Void) {
         self.paths = resolveParamsFiles(args: paths)
         self.action = action
         self.useSTDIN = useSTDIN
@@ -66,13 +69,16 @@ struct LintableFilesVisitor {
         self.parallel = parallel
         self.mode = .lint
         self.allowZeroLintableFiles = allowZeroLintableFiles
+        self.stableGitRevision = stableGitRevision
+        self.explicitConfigurationPaths = explicitConfigurationPaths
         self.block = block
     }
 
     private init(paths: [String], action: String, useSTDIN: Bool, quiet: Bool,
                  useScriptInputFiles: Bool, forceExclude: Bool, useExcludingByPrefix: Bool,
                  cache: LinterCache?, compilerInvocations: CompilerInvocations?,
-                 allowZeroLintableFiles: Bool, block: @escaping (CollectedLinter) -> Void) {
+                 allowZeroLintableFiles: Bool, stableGitRevision: String?, explicitConfigurationPaths: [String],
+                 block: @escaping (CollectedLinter) -> Void) {
         self.paths = resolveParamsFiles(args: paths)
         self.action = action
         self.useSTDIN = useSTDIN
@@ -89,6 +95,8 @@ struct LintableFilesVisitor {
         }
         self.block = block
         self.allowZeroLintableFiles = allowZeroLintableFiles
+        self.stableGitRevision = stableGitRevision
+        self.explicitConfigurationPaths = explicitConfigurationPaths
     }
 
     static func create(_ options: LintOrAnalyzeOptions,
@@ -116,7 +124,10 @@ struct LintableFilesVisitor {
                                                useExcludingByPrefix: options.useExcludingByPrefix,
                                                cache: cache,
                                                compilerInvocations: compilerInvocations,
-                                               allowZeroLintableFiles: allowZeroLintableFiles, block: block)
+                                               allowZeroLintableFiles: allowZeroLintableFiles,
+                                               stableGitRevision: options.stableGitRevision,
+                                               explicitConfigurationPaths: options.configurationFiles,
+                                               block: block)
             return .success(visitor)
         }
     }